Refactor to clean up index.js
1. Create routes/characters.js file to house /characters and
/characters/:id

2. Create utils/helpers file to house generateTs and generateHash
functions

3. Create utils/axiosRequest file to house all main axios methods

4. Create utils/constants file to house marvelUrl
