+++ src/utils/axiosRequest.js
@@ -1,0 +1,9 @@
+const axios = require('axios');
+
+const fetchData = async (url) => {
+  const response = await axios.get(url);
+
+  return response.data;
+};
+
+module.exports = { fetchData };
