+++ src/utils/constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+  marvelUrl: 'http://gateway.marvel.com',
+};
